renamed no avatar image for players for easier reading
